requirements.txt
absl-py==0.7.1
dask==1.2.2
ipython==7.5.0
numpy==1.16.3
pandas==0.24.2
scikit-learn==0.21.2
scipy==1.3.0
sklearn==0.0
toolz==0.9.0
tqdm==4.32.1
